+++ src/Interfaces/http/api/users/index.ts
@@ -5,7 +5,7 @@ import { container } from 'tsyringe';
 import AddUserUseCase from '../../../../Applications/use_cases/AddUserUseCase';
 import { routes } from './routes';
 
-export const usersPlugin: Hapi.Plugin<any> = {
+export const users: Hapi.Plugin<any> = {
   name: "users",
   register: async (server) => {
     const handler = new UserHandler(container.resolve(AddUserUseCase));
